feat(plugins): enable which-key.nvim and adjust config
Enabled the `folke/which-key.nvim` plugin with a preset configuration
for better keybinding hints. Updated the configuration to include
custom icons based on the availability of nerd fonts.

chore(mini): comment out unused clue module

Commented out the `custom.mini.ui.clue` module in the mini core
configuration as it is currently unused.